pyproject.toml
[tool.poetry]
name = "migratable-aws-stacker"
version = "0.1.0"
description = ""
authors = []
packages = [
    { include = "migratable_aws_stacker" }
]

[tool.poetry.dependencies]
python = "^3.9"
stacker_blueprints = "^1.0.8"
aws-cdk-lib = "^2.0.0"
constructs = "^10.0.0"

[tool.poetry.group.dev.dependencies]
nose = "^1.3"
mock = "^2.0"
flake8 = "^3.5"
setuptools = "^69.0.0"
